sigle_gate classにresource preparationメソッドを移動
